+++ rs/cli/src/qualification/retire_blessed_versions.rs
@@ -2,12 +2,9 @@ use backon::{ExponentialBuilder, Retryable};
 use comfy_table::CellAlignment;
 use itertools::Itertools;
 
-use crate::{
-    ctx::DreContext,
-    ic_admin::{ProposeCommand, ProposeOptions},
-};
+use crate::ic_admin::{ProposeCommand, ProposeOptions};
 
-use super::{comfy_table_util::Table, print_table, print_text, Step};
+use super::{comfy_table_util::Table, step::Step, util::StepCtx};
 
 pub struct RetireBlessedVersions {
     pub versions: Vec<String>,
@@ -22,8 +19,8 @@ impl Step for RetireBlessedVersions {
         "retire_blessed_versions".to_string()
     }
 
-    async fn execute(&self, ctx: &DreContext) -> anyhow::Result<()> {
-        let registry = ctx.registry().await;
+    async fn execute(&self, ctx: &StepCtx) -> anyhow::Result<()> {
+        let registry = ctx.dre_ctx().registry().await;
 
         let blessed_versions = registry.elected_guestos()?;
         let mut to_unelect = vec![];
@@ -33,12 +30,13 @@ impl Step for RetireBlessedVersions {
             }
         }
         if to_unelect.is_empty() {
-            print_text(format!("Versions {} are not blessed, skipping step", self.versions.iter().join(",")));
+            ctx.print_text(format!("Versions {} are not blessed, skipping step", self.versions.iter().join(",")));
             return Ok(());
         }
 
         let place_proposal = || async {
-            ctx.ic_admin()
+            ctx.dre_ctx()
+                .ic_admin()
                 .propose_run(
                     ProposeCommand::ReviseElectedVersions {
                         release_artifact: ic_management_types::Artifact::GuestOs,
@@ -65,7 +63,7 @@ impl Step for RetireBlessedVersions {
             .with_rows(blessed_versions.iter().map(|ver| vec![ver.to_string()]).collect_vec())
             .to_table();
 
-        print_table(table);
+        ctx.print_table(table);
 
         Ok(())
     }
