+++ rs/cli/src/qualification/run_xnet_test.rs
@@ -4,7 +4,7 @@ use ic_registry_subnet_type::SubnetType;
 use itertools::Itertools;
 use tokio::process::Command;
 
-use super::{download_canister, download_executable, print_text, Step};
+use super::{step::Step, util::StepCtx};
 
 const E2E_TEST_DRIVER: &str = "e2e-test-driver";
 const XNET_TEST_CANISTER: &str = "xnet-test-canister";
@@ -30,7 +30,7 @@ impl Step for RunXnetTest {
         "xnet_test".to_string()
     }
 
-    async fn execute(&self, ctx: &crate::ctx::DreContext) -> anyhow::Result<()> {
+    async fn execute(&self, ctx: &StepCtx) -> anyhow::Result<()> {
         let key = dirs::home_dir()
             .ok_or(anyhow::anyhow!("Cannot get home directory"))?
             .join(XNET_PRINCIPAL_PATH);
@@ -41,10 +41,10 @@ impl Step for RunXnetTest {
         let file = std::fs::File::open(&key)?;
         file.set_permissions(PermissionsExt::from_mode(0o400))?;
 
-        let e2e_bin = download_executable(E2E_TEST_DRIVER, &self.version).await?;
-        let wasm_path = download_canister(XNET_TEST_CANISTER, &self.version).await?;
+        let e2e_bin = ctx.download_executable(E2E_TEST_DRIVER, &self.version).await?;
+        let wasm_path = ctx.download_canister(XNET_TEST_CANISTER, &self.version).await?;
 
-        let registry = ctx.registry().await;
+        let registry = ctx.dre_ctx().registry().await;
         let subnet = registry.subnets().await?;
         let subnet = subnet
             .values()
@@ -71,7 +71,7 @@ impl Step for RunXnetTest {
             XNET_TEST_NUMBER.to_string(),
         ];
 
-        print_text(format!(
+        ctx.print_text(format!(
             "Running command: XNET_TEST_CANISTER_WASM_PATH={} {} {}",
             wasm_path.display(),
             e2e_bin.display(),
